Update to use Spring profiles
