--- src/main/java/gov/nasa/podaac/swodlr/Environment.java
@@ -1,3 +1,0 @@
-package gov.nasa.podaac.swodlr;
-
-public enum Environment { DEV, PROD }
