+++ src/main/java/gov/nasa/podaac/swodlr/security/config/DevelopmentWebSecurityConfig.java
@@ -1,0 +1,28 @@
+package gov.nasa.podaac.swodlr.security.config;
+
+import java.util.Collections;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.http.HttpMethod;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.reactive.CorsConfigurationSource;
+import org.springframework.web.cors.reactive.UrlBasedCorsConfigurationSource;
+
+@Configuration
+@Profile("dev")
+public class DevelopmentWebSecurityConfig {
+  @Bean
+  public CorsConfigurationSource corsConfigurationSource() {
+    CorsConfiguration config = new CorsConfiguration();
+    config.addAllowedMethod(HttpMethod.GET);
+    config.addAllowedMethod(HttpMethod.POST);
+    config.addAllowedMethod(HttpMethod.HEAD);
+    config.setAllowCredentials(true);
+    config.setAllowedOriginPatterns(Collections.singletonList("*"));
+
+    UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+    source.registerCorsConfiguration("/**", config);
+    return source;
+  }
+}
