+++ src/main/java/gov/nasa/podaac/swodlr/security/config/WebSecurityConfig.java
@@ -1,15 +1,11 @@
-package gov.nasa.podaac.swodlr.security;
+package gov.nasa.podaac.swodlr.security.config;
 
-import gov.nasa.podaac.swodlr.Environment;
-import gov.nasa.podaac.swodlr.SwodlrProperties;
 import gov.nasa.podaac.swodlr.security.authentication.client.JweCookieReactiveOauth2AuthorizedClientService;
 import gov.nasa.podaac.swodlr.security.authentication.handlers.SuccessMessageAuthenticationSuccessHandler;
 import gov.nasa.podaac.swodlr.security.authentication.handlers.UserBootstrapAuthenticationSuccessHandler;
-import java.util.Collections;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Profile;
-import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.web.reactive.EnableWebFluxSecurity;
 import org.springframework.security.config.web.server.ServerHttpSecurity;
 import org.springframework.security.oauth2.client.AuthorizedClientServiceReactiveOAuth2AuthorizedClientManager;
@@ -18,9 +14,6 @@ import org.springframework.security.oauth2.client.ReactiveOAuth2AuthorizedClient
 import org.springframework.security.oauth2.client.registration.ReactiveClientRegistrationRepository;
 import org.springframework.security.web.server.SecurityWebFilterChain;
 import org.springframework.security.web.server.authentication.DelegatingServerAuthenticationSuccessHandler;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.reactive.CorsConfigurationSource;
-import org.springframework.web.cors.reactive.UrlBasedCorsConfigurationSource;
 
 @EnableWebFluxSecurity
 @Profile({"!test"})
@@ -33,9 +26,6 @@ public class WebSecurityConfig {
 
   @Autowired
   private SuccessMessageAuthenticationSuccessHandler successHandler;
-
-  @Autowired
-  private SwodlrProperties swodlrProperties;
 
   public WebSecurityConfig(ReactiveClientRegistrationRepository clientRegistrationRepository) {
     authorizedClientService = new JweCookieReactiveOauth2AuthorizedClientService();
@@ -72,22 +62,5 @@ public class WebSecurityConfig {
   @Bean
   public ReactiveOAuth2AuthorizedClientService authorizedClientService() {
     return authorizedClientService;
-  }
-
-  @Bean
-  public CorsConfigurationSource corsConfigurationSource() {
-    CorsConfiguration config = new CorsConfiguration();
-    
-    if (swodlrProperties.getEnv() == Environment.DEV) {
-      config.addAllowedMethod(HttpMethod.GET);
-      config.addAllowedMethod(HttpMethod.POST);
-      config.addAllowedMethod(HttpMethod.HEAD);
-      config.setAllowCredentials(true);
-      config.setAllowedOriginPatterns(Collections.singletonList("*"));
-    }
-
-    UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-    source.registerCorsConfiguration("/**", config);
-    return source;
   }
 }
