+++ terraform/environments/sit.env
@@ -2,4 +2,4 @@ export REGION=us-west-2
 export BUCKET=podaac-services-sit-terraform
 
 export TF_VAR_container_image_tag=develop
-export TF_VAR_environment=DEV
+export TF_VAR_active_profiles="['dev']"
